[Unified observability] Wrap exploratory view in the page component (#124269)
(cherry picked from commit ec47913330eb1243729fd78adb21a47158354a07)
