+++ x-pack/plugins/observability/public/components/shared/exploratory_view/exploratory_view.test.tsx
@@ -50,8 +50,6 @@ describe('ExploratoryView', () => {
 
     expect(await screen.findByText(/No series found. Please add a series./i)).toBeInTheDocument();
     expect(await screen.findByText(/Hide chart/i)).toBeInTheDocument();
-    expect(await screen.findByText(/Refresh/i)).toBeInTheDocument();
-    expect(await screen.findByRole('heading', { name: /Explore data/i })).toBeInTheDocument();
   });
 
   it('renders lens component when there is series', async () => {
@@ -61,5 +59,18 @@ describe('ExploratoryView', () => {
     expect(await screen.findByText(/Lens Embeddable Component/i)).toBeInTheDocument();
 
     expect(screen.getByTestId('exploratoryViewSeriesPanel0')).toBeInTheDocument();
+  });
+
+  it('shows/hides the chart', async () => {
+    render(<ExploratoryView />);
+    expect(screen.queryByText('Refresh')).toBeInTheDocument();
+
+    const toggleButton = await screen.findByText('Hide chart');
+    expect(toggleButton).toBeInTheDocument();
+
+    toggleButton.click();
+
+    expect(toggleButton.textContent).toBe('Show chart');
+    expect(screen.queryByText('Refresh')).toBeNull();
   });
 });
