+++ x-pack/plugins/observability/public/components/shared/exploratory_view/exploratory_view.tsx
@@ -9,17 +9,16 @@ import { i18n } from '@kbn/i18n';
 import React, { useEffect, useRef, useState } from 'react';
 import styled from 'styled-components';
 import {
+  EuiButton,
   EuiButtonEmpty,
   EuiResizableContainer,
   EuiTitle,
-  EuiPanel,
   EuiFlexGroup,
   EuiFlexItem,
 } from '@elastic/eui';
 import { PanelDirection } from '@elastic/eui/src/components/resizable_container/types';
 import { useKibana } from '../../../../../../../src/plugins/kibana_react/public';
 import { ObservabilityPublicPluginsStart } from '../../../plugin';
-import { ExploratoryViewHeader } from './header/header';
 import { useSeriesStorage } from './hooks/use_series_storage';
 import { useLensAttributes } from './hooks/use_lens_attributes';
 import { TypedLensByValueInput } from '../../../../../lens/public';
@@ -27,7 +26,9 @@ import { useAppIndexPatternContext } from './hooks/use_app_index_pattern';
 import { SeriesViews } from './views/series_views';
 import { LensEmbeddable } from './lens_embeddable';
 import { EmptyView } from './components/empty_view';
-import type { ChartTimeRange } from './header/last_updated';
+import { ChartTimeRange, LastUpdated } from './header/last_updated';
+import { useExpViewTimeRange } from './hooks/use_time_range';
+import { ExpViewActionMenu } from './components/action_menu';
 
 export type PanelId = 'seriesPanel' | 'chartPanel';
 
@@ -53,9 +54,10 @@ export function ExploratoryView({
 
   const { loadIndexPattern, loading } = useAppIndexPatternContext();
 
-  const { firstSeries, allSeries, lastRefresh, reportType } = useSeriesStorage();
+  const { firstSeries, allSeries, lastRefresh, reportType, setLastRefresh } = useSeriesStorage();
 
   const lensAttributesT = useLensAttributes();
+  const timeRange = useExpViewTimeRange();
 
   const setHeightOffset = () => {
     if (seriesBuilderRef?.current && wrapperRef.current) {
@@ -100,80 +102,85 @@ export function ExploratoryView({
     }
   };
 
-  return (
-    <Wrapper>
-      {lens ? (
-        <>
-          <ExploratoryViewHeader
-            lensAttributes={lensAttributes}
-            chartTimeRange={chartTimeRangeContext}
-          />
-          <LensWrapper ref={wrapperRef} height={height}>
-            <ResizableContainer direction="vertical" onToggleCollapsed={onCollapse}>
-              {(EuiResizablePanel, EuiResizableButton, { togglePanel }) => {
-                collapseFn.current = (id, direction) => togglePanel?.(id, { direction });
+  return lens ? (
+    <>
+      <ExpViewActionMenu timeRange={timeRange} lensAttributes={lensAttributes} />
+      <LensWrapper ref={wrapperRef} height={height}>
+        <ResizableContainer direction="vertical" onToggleCollapsed={onCollapse}>
+          {(EuiResizablePanel, _EuiResizableButton, { togglePanel }) => {
+            collapseFn.current = (id, direction) => togglePanel?.(id, { direction });
 
-                return (
-                  <>
-                    <EuiFlexGroup alignItems="flexStart">
-                      <EuiFlexItem grow={false}>
-                        <EuiButtonEmpty
-                          {...(hiddenPanel === 'chartPanel'
-                            ? { iconType: 'arrowRight' }
-                            : { iconType: 'arrowDown' })}
-                          onClick={() => onChange('chartPanel')}
-                        >
-                          {hiddenPanel === 'chartPanel' ? SHOW_CHART_LABEL : HIDE_CHART_LABEL}
-                        </EuiButtonEmpty>
+            return (
+              <>
+                <EuiFlexGroup alignItems="center">
+                  <EuiFlexItem grow={false}>
+                    <EuiButtonEmpty
+                      {...(hiddenPanel === 'chartPanel'
+                        ? { iconType: 'arrowRight' }
+                        : { iconType: 'arrowDown' })}
+                      onClick={() => onChange('chartPanel')}
+                    >
+                      {hiddenPanel === 'chartPanel' ? SHOW_CHART_LABEL : HIDE_CHART_LABEL}
+                    </EuiButtonEmpty>
+                  </EuiFlexItem>
+                  {hiddenPanel === 'chartPanel' ? null : (
+                    <>
+                      <EuiFlexItem style={{ textAlign: 'right' }}>
+                        <LastUpdated chartTimeRange={chartTimeRangeContext} />
                       </EuiFlexItem>
-                    </EuiFlexGroup>
+                      <EuiFlexItem grow={false}>
+                        <EuiButton iconType="refresh" onClick={() => setLastRefresh(Date.now())}>
+                          {REFRESH_LABEL}
+                        </EuiButton>
+                      </EuiFlexItem>
+                    </>
+                  )}
+                </EuiFlexGroup>
 
-                    <EuiResizablePanel
-                      initialSize={40}
-                      minSize={'30%'}
-                      mode={'collapsible'}
-                      id="chartPanel"
-                    >
-                      {lensAttributes ? (
-                        <LensEmbeddable
-                          setChartTimeRangeContext={setChartTimeRangeContext}
-                          lensAttributes={lensAttributes}
-                        />
-                      ) : (
-                        <EmptyView series={firstSeries} loading={loading} reportType={reportType} />
-                      )}
-                    </EuiResizablePanel>
+                <EuiResizablePanel
+                  initialSize={40}
+                  minSize={'30%'}
+                  mode={'collapsible'}
+                  id="chartPanel"
+                >
+                  {lensAttributes ? (
+                    <LensEmbeddable
+                      setChartTimeRangeContext={setChartTimeRangeContext}
+                      lensAttributes={lensAttributes}
+                    />
+                  ) : (
+                    <EmptyView series={firstSeries} loading={loading} reportType={reportType} />
+                  )}
+                </EuiResizablePanel>
 
-                    <EuiResizablePanel
-                      initialSize={60}
-                      minSize="10%"
-                      mode={'main'}
-                      id="seriesPanel"
-                      color="subdued"
-                      className="paddingTopSmall"
-                    >
-                      <SeriesViews
-                        seriesBuilderRef={seriesBuilderRef}
-                        onSeriesPanelCollapse={onChange}
-                      />
-                    </EuiResizablePanel>
-                  </>
-                );
-              }}
-            </ResizableContainer>
-            {hiddenPanel === 'seriesPanel' && (
-              <ShowPreview onClick={() => onChange('seriesPanel')} iconType="arrowUp">
-                {PREVIEW_LABEL}
-              </ShowPreview>
-            )}
-          </LensWrapper>
-        </>
-      ) : (
-        <EuiTitle>
-          <h2>{LENS_NOT_AVAILABLE}</h2>
-        </EuiTitle>
-      )}
-    </Wrapper>
+                <EuiResizablePanel
+                  initialSize={60}
+                  minSize="10%"
+                  mode={'main'}
+                  id="seriesPanel"
+                  color="subdued"
+                  className="paddingTopSmall"
+                >
+                  <SeriesViews
+                    seriesBuilderRef={seriesBuilderRef}
+                    onSeriesPanelCollapse={onChange}
+                  />
+                </EuiResizablePanel>
+              </>
+            );
+          }}
+        </ResizableContainer>
+        {hiddenPanel === 'seriesPanel' && (
+          <ShowPreview onClick={() => onChange('seriesPanel')} iconType="arrowUp">
+            {PREVIEW_LABEL}
+          </ShowPreview>
+        )}
+      </LensWrapper>
+    </>
+  ) : (
+    <EuiTitle>
+      <h2>{LENS_NOT_AVAILABLE}</h2>
+    </EuiTitle>
   );
 }
 const LensWrapper = styled.div<{ height: string }>`
@@ -192,19 +199,6 @@ const ResizableContainer = styled(EuiResizableContainer)`
   }
 `;
 
-const Wrapper = styled(EuiPanel)`
-  max-width: 1800px;
-  min-width: 800px;
-  margin: 0 auto;
-  width: 100%;
-  overflow-x: auto;
-  position: relative;
-
-  .echLegendItem__action {
-    display: none;
-  }
-`;
-
 const ShowPreview = styled(EuiButtonEmpty)`
   position: absolute;
   bottom: 34px;
@@ -220,6 +214,10 @@ const SHOW_CHART_LABEL = i18n.translate('xpack.observability.overview.explorator
 
 const PREVIEW_LABEL = i18n.translate('xpack.observability.overview.exploratoryView.preview', {
   defaultMessage: 'Preview',
+});
+
+const REFRESH_LABEL = i18n.translate('xpack.observability.overview.exploratoryView.refresh', {
+  defaultMessage: 'Refresh',
 });
 
 const LENS_NOT_AVAILABLE = i18n.translate(
