+++ x-pack/plugins/observability/public/components/shared/exploratory_view/index.tsx
@@ -10,7 +10,6 @@ import { i18n } from '@kbn/i18n';
 import { useHistory } from 'react-router-dom';
 import { ExploratoryView } from './exploratory_view';
 import { useKibana } from '../../../../../../../src/plugins/kibana_react/public';
-import { euiStyled } from '../../../../../../../src/plugins/kibana_react/common';
 import { ObservabilityPublicPluginsStart } from '../../../plugin';
 import { useBreadcrumbs } from '../../../hooks/use_breadcrumbs';
 import { IndexPatternContextProvider } from './hooks/use_app_index_pattern';
@@ -22,6 +21,11 @@ import {
 import { UrlStorageContextProvider } from './hooks/use_series_storage';
 import { useTrackPageview } from '../../..';
 import { TypedLensByValueInput } from '../../../../../lens/public';
+import { usePluginContext } from '../../../hooks/use_plugin_context';
+
+const PAGE_TITLE = i18n.translate('xpack.observability.expView.heading.label', {
+  defaultMessage: 'Explore data',
+});
 
 export interface ExploratoryViewPageProps {
   useSessionStorage?: boolean;
@@ -52,6 +56,7 @@ export function ExploratoryViewPage({
     app
   );
 
+  const { ObservabilityPageTemplate } = usePluginContext();
   const {
     services: { uiSettings, notifications },
   } = useKibana<ObservabilityPublicPluginsStart>();
@@ -67,19 +72,15 @@ export function ExploratoryViewPage({
       });
 
   return (
-    <Wrapper>
+    <ObservabilityPageTemplate pageHeader={{ pageTitle: PAGE_TITLE }}>
       <IndexPatternContextProvider>
         <UrlStorageContextProvider storage={kbnUrlStateStorage}>
           <ExploratoryView saveAttributes={saveAttributes} />
         </UrlStorageContextProvider>
       </IndexPatternContextProvider>
-    </Wrapper>
+    </ObservabilityPageTemplate>
   );
 }
-
-const Wrapper = euiStyled.div`
-  padding: ${(props) => props.theme.eui.paddingSizes.l};
-`;
 
 // eslint-disable-next-line import/no-default-export
 export default ExploratoryViewPage;
